Fixed NED tile downloader to handle file names with different formats
The script first retrieves a list of all available files in the folder, then indexes them by the lat/lon key (i.e. "n40w120").  Since the files in the zips can be named differently also, the script then looks for the first IMG file in the zip and extracts it.  Finally the script renames the IMG file to a simple format based on the lat/lon key (i.e. "n40w120.img").

The regex used to search IMG files in the zip is probably overkill.