Adding example field boundaries
The primary purpose of the field boundaries in the example is to demonstrate constraining the location of the calibration points.  It may be necessary to add additional fields for the wildlife refuge and/or remove some of the fields in this dataset since it was developed for 2016 but the example is 2015.