Added "fmask_smooth_flag" and "fmask_smooth_cells" parameters to landsat_prep_scene
The smoothing follows an approach similar to the ArcGIS "Boundary Clean" tool.  The cloud pixels are eroded/dilated and then dilated/eroded to help smooth out the boundaries.  Originally the erode section was eroding and then dilating, but the buffer section was only dilating, which seemed confusing.  Now the erode and buffer sections only do the eroding and dilating that is requested.

I set the default values if not set to False for all three parameters.