Fixing issues with landsat_prep_scene "smooth" parameter
The purpose of this parameter was to remove the fringe artifacts at the edge of most Landsat 5 and some Landsat 7 images.  It worked well for that but it was causing some weird issues with the cloud masks.  First, the parameter is now set in the project INI, instead of being a separate command line parameter.  Second, the parameter uses different buffer sizes for Landsat 5 and 7 sicne the fringes are shaped differently.  The default is still set to True, but this may need to be revisited.