Fixed bug in WRS2 download script
The URL for the WRS2 shapefile changed slightly (WRS is now uppercase in the file name).  I think it would be better to change all the INIs to expect the uppercase version of the filename, but for now, renaming the shapefile to lowercase should minimize disruption in the current workflow.  See issue #38 for more details.  This may be something we want to revisit in a 0.2.0 release.